Refactor computed swaps
