+++ packages/frontend/package.json
@@ -39,6 +39,7 @@
         "bnc-onboard": "1.27.0",
         "chart.js": "^2.9.4",
         "clsx": "^1.1.1",
+        "constate": "^3.3.2",
         "copy-to-clipboard": "^3.3.1",
         "date-fns": "^2.28.0",
         "fathom-client": "^3.2.0",
